completed the Data Engineer test
kicking myself to be honest, question 1 I got 100% on, question 2 was a repeat from the Software Dev test so I tried to improve my answer, and then somehow I made it break? I really don't know what happened but it got a 25% score when I submitted it. Question 3 was funky because I didn't realize SQL was on this, so I had to get copilot to explain a lot of concepts, but luckily the question was simple and I got 100% somehow! my overall was 75% which is better than yesterdays, though of course I'm less happy with it :D. results are at this link for the next month: https://app.codility.com/c/feedback/4AM773-S4V/